GraphView refactoring

+++ Graphene/Source/graphene/Structure.hpp
@@ -175,7 +175,7 @@ namespace Graphene {
 
     public:
         Vec2f center = Vec2f(0.0f, 0.0f);
-        float radius = 0.0f;
+        double radius = 0.0f;
         std::pair<Vec2f, Vec2f> bb = {{0.0f, 0.0f}, {0.0f, 0.0f}};
         std::pair<Vec2f, Vec2f> bbBack = {{0.0f, 0.0f}, {0.0f, 0.0f}};
